라우터 및 컴포넌트 구조 변경
