+++ src/main/front/src/pages/index.jsx
@@ -1,0 +1,3 @@
+export { default as Login } from '../security/containers/LoginContainer';
+export { default as Join } from '../security/containers/JoinContainer'
+export { default as NotFount } from '../common/components/NotFoundComponent'
