+++ src/main/front/src/security/components/JoinComponent.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import {Typography} from 'antd';
+import 'antd/dist/antd.css';  // or 'antd/dist/antd.less'
+
+
+const JoinComponent = () => {
+    const { Title } = Typography;
+
+    return (
+        <Title>회원 가입</Title>
+    );
+}
+
+export default JoinComponent;
