--- src/main/front/src/security/routers/join.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Route } from "react-router-dom";
-import JoinContainer from '../containers/JoinContainer';
-
-
-const Join = () => {
-    return (
-        <Route path="/join" component={JoinContainer} />
-    );
-}
-
-export default Join;
